requirements.txt
streamlit
sqlalchemy
psycopg2-binary
pandas
openai
langchain
python-dotenv
langchain-openai
fastapi
uvicorn
flask